feat(grouping): #MVS-51 create, update and populate grouping is now possible
