+++ src/main/java/fr/openent/viescolaire/security/Grouping/GroupAndClassManage.java
@@ -1,8 +1,9 @@
-package fr.openent.viescolaire.security;
+package fr.openent.viescolaire.security.Grouping;
 
 import fr.openent.Viescolaire;
 import fr.openent.viescolaire.core.constants.Field;
 import fr.openent.viescolaire.helper.PromiseHelper;
+import fr.openent.viescolaire.security.WorkflowActionUtils;
 import fr.wseduc.webutils.http.Binding;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
@@ -16,29 +17,36 @@ import org.entcore.common.sql.Sql;
 import org.entcore.common.sql.SqlResult;
 import org.entcore.common.user.UserInfos;
 
-public class GroupingRights implements ResourcesProvider {
-    private static final Logger log = LoggerFactory.getLogger(GroupingRights.class);
+public class GroupAndClassManage implements ResourcesProvider {
+    private static final Logger log = LoggerFactory.getLogger(GroupAndClassManage.class);
 
     @Override
     public void authorize(final HttpServerRequest resourceRequest, Binding binding, final UserInfos user, final Handler<Boolean> handler) {
         String id = resourceRequest.getParam(Field.ID);
         checkGroupingsRights(user, id)
-                .onSuccess(res -> handler.handle(res && WorkflowActionUtils.hasRight(user, WorkflowActionUtils.ADMIN_RIGHT)))
-                .onFailure(err -> {
-                    handler.handle(false);
-                });
+                .onSuccess(res -> {
+                    handler.handle(res && WorkflowActionUtils.hasRight(user, WorkflowActionUtils.ADMIN_RIGHT)
+                            && (user.getGroupsIds().contains(resourceRequest.getParam(Field.STUDENT_DIVISION_ID))
+                            || user.getClasses().contains(resourceRequest.getParam(Field.STUDENT_DIVISION_ID))));
+                })
+                .onFailure(err -> handler.handle(false));
     }
 
+    /**
+     * Check if the user have the rights to manage a specific grouping.
+     * @param user          User data
+     * @param groupingId    Identifier of the grouping
+     * @return              Return a future with the result of the check
+     */
     private Future<Boolean> checkGroupingsRights(UserInfos user, String groupingId) {
         Promise<Boolean> promise = Promise.promise();
         String query = "SELECT structure_id FROM " + Viescolaire.VSCO_SCHEMA + "." + Viescolaire.GROUPING_TABLE + " WHERE id = ? ;";
         JsonArray values = new JsonArray();
         values.add(groupingId);
         Sql.getInstance().prepared(query, values, SqlResult.validUniqueResultHandler(res -> {
-            if(res.isRight() && !res.right().getValue().isEmpty()){
+            if (res.isRight() && !res.right().getValue().isEmpty()) {
                 promise.complete(user.getStructures().contains(res.right().getValue().getString(Field.STRUCTURE_ID)));
-            }
-            else {
+            } else {
                 String messageToFormat = "[Viescolaire@%s::isUserAllowToManageGroupings] Error while checking rights : %s";
                 PromiseHelper.reject(log, messageToFormat, this.getClass().getSimpleName(), new Exception(res.left().getValue()), promise);
             }
