+++ src/main/java/fr/openent/viescolaire/security/Grouping/StructureOwnerFilter.java
@@ -1,6 +1,7 @@
-package fr.openent.viescolaire.security;
+package fr.openent.viescolaire.security.Grouping;
 
 import fr.openent.viescolaire.core.constants.Field;
+import fr.openent.viescolaire.security.WorkflowActionUtils;
 import fr.wseduc.webutils.http.Binding;
 import io.vertx.core.Handler;
 import io.vertx.core.http.HttpServerRequest;
